+++ Common/src/main/java/games/alejandrocoria/mapfrontiers/client/gui/component/SimpleSlider.java
@@ -1,7 +1,7 @@
 package games.alejandrocoria.mapfrontiers.client.gui.component;
 
 import games.alejandrocoria.mapfrontiers.client.gui.ColorConstants;
-import games.alejandrocoria.mapfrontiers.client.mixin.AbstractSliderButtonAccessor;
+import games.alejandrocoria.mapfrontiers.mixin.client.AbstractSliderButtonAccessor;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractSliderButton;