+++ doc/changelog.md
@@ -1,6 +1,7 @@
 **Beta 4:**
 * Added: Frontier info paste options now support copying Path style between Path frontiers.
 * Fixed: Changing frontier info paste options no longer sends unnecessary frontier updates to the server.
+* Fixed: OP permission changes from the server only applying after the player disconnected and reconnected.
 
 **Beta 3:**
 * Added: More colors to the color palette.